+++ code/engine/system.h
@@ -25,7 +25,7 @@
 ENGINE_API extern FLOAT GetTime();
 ENGINE_API extern FLOAT ScaleBetween(FLOAT x, FLOAT a, FLOAT b, FLOAT na, FLOAT nb);
 
-extern size_t gMemUsed, gMemUsedLua, gMemPeak, gResourceCount;
+extern UINT64 gMemUsed, gMemUsedLua, gMemPeak, gResourceCount;
 extern VOID neon_mempeak_update();
 extern LPVOID neon_malloc(size_t size);
 extern LPVOID neon_realloc(LPVOID mem, size_t newSize);
